requirements.txt
SpeechRecognition
numpy
pandas
streamlit
python-dotenv
pyaudio
groq